+++ examples/object_detection_3d/pyproject.toml
@@ -15,6 +15,7 @@ openmim = "^0.3.9"
 mmcv-lite = ">=2.0.0rc4"
 mmdet = ">=3.0.0"
 mmdet3d = ">=1.1.0"
+dacite = "^1.8.1"
 
 [tool.poetry.group.dev.dependencies]
 pre-commit = "^2.17"
